Move Helper Operators & Extra Links to airflow/utils/tests.py
This are moved because they are needed by both plugin_manager.py (as it needs to import using imp.load_source when loading test_plugin.py) and  test_views.py.
